requirements.txt
numpy==1.19.5
Pillow==8.1.0
